requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask==3.0.0
APScheduler==3.10.4
python-dotenv==1.0.0
alembic==1.13.1
Jinja2==3.1.2
gunicorn==21.2.0 